+++ cognite/client/data_classes/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from cognite.client.data_classes._base import Geometry
+from cognite.client.data_classes.aggregations import CountAggregate
 from cognite.client.data_classes.annotations import (
     Annotation,
     AnnotationFilter,
@@ -27,7 +28,6 @@ from cognite.client.data_classes.contextualization import (
 )
 from cognite.client.data_classes.data_sets import (
     DataSet,
-    DataSetAggregate,
     DataSetFilter,
     DataSetList,
     DataSetUpdate,
@@ -67,7 +67,6 @@ from cognite.client.data_classes.extractionpipelines import (
     ExtractionPipelineUpdate,
 )
 from cognite.client.data_classes.files import (
-    FileAggregate,
     FileMetadata,
     FileMetadataFilter,
     FileMetadataList,
@@ -123,7 +122,6 @@ from cognite.client.data_classes.relationships import (
 )
 from cognite.client.data_classes.sequences import (
     Sequence,
-    SequenceAggregate,
     SequenceColumn,
     SequenceColumnList,
     SequenceColumnUpdate,
@@ -172,7 +170,6 @@ from cognite.client.data_classes.three_d import (
 )
 from cognite.client.data_classes.time_series import (
     TimeSeries,
-    TimeSeriesAggregate,
     TimeSeriesFilter,
     TimeSeriesList,
     TimeSeriesUpdate,
@@ -243,6 +240,7 @@ __all__ = [
     "AnnotationUpdate",
     "AggregateResultItem",
     "Asset",
+    "CountAggregate",
     "AssetAggregate",
     "AssetFilter",
     "AssetHierarchy",
@@ -276,7 +274,6 @@ __all__ = [
     "EventFilter",
     "EventList",
     "EventUpdate",
-    "FileAggregate",
     "FileMetadata",
     "FileMetadataFilter",
     "FileMetadataList",
@@ -304,7 +301,6 @@ __all__ = [
     "RelationshipList",
     "RelationshipUpdate",
     "Sequence",
-    "SequenceAggregate",
     "SequenceColumnUpdate",
     "SequenceData",
     "SequenceRows",
@@ -339,7 +335,6 @@ __all__ = [
     "ThreeDNode",
     "ThreeDNodeList",
     "TimeSeries",
-    "TimeSeriesAggregate",
     "TimeSeriesFilter",
     "TimeSeriesList",
     "TimeSeriesUpdate",
@@ -369,7 +364,6 @@ __all__ = [
     "TransformationSchemaColumn",
     "TransformationSchemaColumnList",
     "DataSet",
-    "DataSetAggregate",
     "DataSetFilter",
     "DataSetList",
     "DataSetUpdate",