+++ cognite/client/data_classes/files.py
@@ -8,7 +8,6 @@ from cognite.client.data_classes._base import (
     CogniteListUpdate,
     CogniteObjectUpdate,
     CognitePrimitiveUpdate,
-    CognitePropertyClassUtil,
     CogniteResource,
     CogniteResourceList,
     CogniteUpdate,
@@ -95,8 +94,8 @@ class FileMetadata(CogniteResource):
     def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> FileMetadata:
         instance = super()._load(resource, cognite_client)
         instance.labels = Label._load_list(instance.labels)
-        if instance.geo_location is not None:
-            instance.geo_location = GeoLocation.load(instance.geo_location)
+        if isinstance(instance.geo_location, dict):
+            instance.geo_location = GeoLocation._load(instance.geo_location)
         return instance
 
     def dump(self, camel_case: bool = False) -> dict[str, Any]:
@@ -125,8 +124,8 @@ class FileMetadataFilter(CogniteFilter):
         created_time (dict[str, Any] | TimestampRange | None): Range between two timestamps.
         last_updated_time (dict[str, Any] | TimestampRange | None): Range between two timestamps.
         uploaded_time (dict[str, Any] | TimestampRange | None): Range between two timestamps.
-        source_created_time (dict[str, Any] | None): Filter for files where the sourceCreatedTime field has been set and is within the specified range.
-        source_modified_time (dict[str, Any] | None): Filter for files where the sourceModifiedTime field has been set and is within the specified range.
+        source_created_time (dict[str, Any] | TimestampRange | None): Filter for files where the sourceCreatedTime field has been set and is within the specified range.
+        source_modified_time (dict[str, Any] | TimestampRange | None): Filter for files where the sourceModifiedTime field has been set and is within the specified range.
         external_id_prefix (str | None): Filter by this (case-sensitive) prefix for the external ID.
         directory_prefix (str | None): Filter by this (case-sensitive) prefix for the directory provided by the client.
         uploaded (bool | None): Whether or not the actual file is uploaded. This field is returned only by the API, it has no effect in a post body.
@@ -147,8 +146,8 @@ class FileMetadataFilter(CogniteFilter):
         created_time: dict[str, Any] | TimestampRange | None = None,
         last_updated_time: dict[str, Any] | TimestampRange | None = None,
         uploaded_time: dict[str, Any] | TimestampRange | None = None,
-        source_created_time: dict[str, Any] | None = None,
-        source_modified_time: dict[str, Any] | None = None,
+        source_created_time: dict[str, Any] | TimestampRange | None = None,
+        source_modified_time: dict[str, Any] | TimestampRange | None = None,
         external_id_prefix: str | None = None,
         directory_prefix: str | None = None,
         uploaded: bool | None = None,
@@ -182,7 +181,15 @@ class FileMetadataFilter(CogniteFilter):
         if isinstance(self.labels, LabelFilter):
             result["labels"] = self.labels.dump(camel_case)
         if isinstance(self.geo_location, GeoLocationFilter):
-            result["geoLocation"] = self.geo_location.dump(camel_case)
+            result["geoLocation" if camel_case else "geo_location"] = self.geo_location.dump(camel_case)
+        keys = (
+            ["createdTime", "lastUpdatedTime", "uploadedTime", "sourceCreatedTime", "sourceModifiedTime"]
+            if camel_case
+            else ["created_time", "last_updated_time", "uploaded_time", "source_created_time", "source_modified_time"]
+        )
+        for key in keys:
+            if key in result and isinstance(result[key], TimestampRange):
+                result[key] = result[key].dump(camel_case)
         return result
 
 
@@ -288,21 +295,6 @@ class FileMetadataUpdate(CogniteUpdate):
             PropertySpec("labels", is_container=True),
             PropertySpec("geo_location"),
         ]
-
-
-class FileAggregate(dict):
-    """Aggregation results for files
-
-    Args:
-        count (int | None): Number of filtered items included in aggregation
-        **kwargs (Any): No description.
-    """
-
-    def __init__(self, count: int | None = None, **kwargs: Any) -> None:
-        self.count = count
-        self.update(kwargs)
-
-    count = CognitePropertyClassUtil.declare_property("count")
 
 
 class FileMetadataList(CogniteResourceList[FileMetadata], IdTransformerMixin):